Get Datasets

This function pulls a list of datasets from TCRM

get_datasets(
  datasetTypes = NULL,
  folderId = NULL,
  hasCurrentOnly = NULL,
  ids = NULL,
  includeCurrentVersion = NULL,
  licenseType = NULL,
  order = NULL,
  page = NULL,
  pageSize = NULL,
  q = NULL,
  scope = NULL,
  sort = NULL,
  supportsNewDates = NULL,
  debug = FALSE,
  consumer_key = Sys.getenv("CONSUMER_KEY"),
  consumer_secret = Sys.getenv("CONSUMER_ID")
)

Arguments

datasetTypes

The type of the dataset. Valid values are: 'Default', 'Live', 'Staged', 'Data'. Optional 50.0

folderId

Filters the results to include only the contents of a specific folder. The id can be the requesting user’s ID for items in the user’s private folder. Optional 36.0

hasCurrentOnly

Boolean Indicates whether to filter the list of datasets to include only those datasets that have a current version (true) or not (false).. The default is false. Optional 52.0

ids

Filter the list of datasets to include only datasets with the specified ids. Optional 53.0

includeCurrentVersion

Boolean Indicates whether to include the current dataset version metadata in the collection (true) or not (false).. The default is false. Optional 52.0

licenseType

Filters the collection by the Analytics license type.. Valid values are: 'EinsteinAnalytics' (CRM Analytics), 'Sonic' (Salesforce Data Pipeline). Optional 52.0

order

Ordering applied to the results. Valid values are: 'Ascending', 'Descending'. Optional 42.0

page

A generated token that indicates the view of datasets to be returned. Optional 36.0

pageSize

The number of items to be returned in a single page. Minimum is 1, maximum is 200, and default is 25. Optional 36.0

q

Search terms. Individual terms are separated by spaces. A wildcard is automatically appended to the last token in the query string. If the user’s search query contains quotation marks or wildcards, those symbols are automatically removed from the query string in the URI along with any other special characters. Optional 36.0

scope

The type of scope to be applied to the returned collection. Valid values are: 'CreatedByMe', 'Mru' (Most Recently Used), 'SharedWithMe'. Optional 38.0

sort

The type of sort order to be applied to the returned collection. Valid values are: 'App', 'CreatedBy', 'CreatedDate', 'LastModified', 'LastModifiedBy', 'Mru' (Most Recently Used, last viewed date), 'Name', 'Type'. Optional 38.0

supportsNewDates

Indicates whether to filters the list of datasets to include only those datasets that support new dates (true) or not (false).. The default is false. Optional 52.0

debug

Outputs the call in the Console for debugging purposes

consumer_key

Consumer key given by CRMA App

consumer_secret

Consumer secret given by CRMA App

Value

Dataframe of datasets and metadata